fix: type total charge as int so external calculators accept it
A system's total charge is physically an integer, but the benchmark input
models (`Molecule`, `Fragment`, `Conformer`, `Reaction`, `TautomerPair`,
`MolecularSystem`, …) and the `STRUCTURE_CHARGES` tables typed/stored it as a
float, and the MD/analysis paths assigned `atoms.info["charge"] = float(...)`.

Native mlip force fields tolerate a float charge, but external ASE calculators
(notably fairchem/UMA's `FAIRChemCalculator`, omol task) reject it with
`Invalid type for charge: <class 'float'>. Charge must be an integer ...`,
which made every molecular benchmark (folding_stability, sampling, etc.) error
out for UMA. This mirrors NEB, which already types `charge: int`.

Fix: type the charge fields and `STRUCTURE_CHARGES` as `int` (pydantic coerces
the integral source values at parse time) and drop the `float(...)` wrappers at
the `atoms.info["charge"] = ...` assignment sites. No behavioural change for
native models; unblocks external calculators.

+++ src/mlipaudit/benchmarks/dihedral_scan/dihedral_scan.py
@@ -63,7 +63,7 @@ class Fragment(BaseModel):
     atom_symbols: list[str]
     conformer_coordinates: list[list[tuple[float, float, float]]]
     smiles: str
-    charge: float = DEFAULT_CHARGE
+    charge: int = DEFAULT_CHARGE
 
 
 Fragments = TypeAdapter(dict[str, Fragment])
@@ -204,7 +204,7 @@ class DihedralScanBenchmark(Benchmark):
         for fragment_name, fragment in self._torsion_net_500.items():
             for conf_coord in fragment.conformer_coordinates:
                 atoms = Atoms(symbols=fragment.atom_symbols, positions=conf_coord)
-                atoms.info["charge"] = float(fragment.charge)
+                atoms.info["charge"] = fragment.charge
                 atoms.info["spin"] = DEFAULT_SPIN
                 atoms_list_all_structures.append(atoms)
                 structure_indices_map[fragment_name].append(index)
